Refactor SearchForSubgraphPattern Code to be more efficient and make use of Node neighbours member variable.
Because:
- Initial code is very inefficient with a large time complexity and waste of resources.

edge iteration loops were removed as they are not necessary if the node neighbours variable is used. The nodeNeighbours arraylist in the node class is changed to an arraylist of nodes instead of integers. this is for better access to node information. The methods that require the original int ids of the nodes were refactored to accommodate having the nodes as the arraylist values.

Signed-off-by: Edan Toledo <TLDEDA001@myuct.ac.za>