Fix DisplaySubset to correctly display all adjacent nodes in graph object.
-Ensures Subsets are constructed correctly and returned to the front-end.

Add DirectedEdgeNeighbours and undirectedEdgeNeighbours, correct the code that used old implementation of neighbour code.

Add Methods in graph to select a subset of the token list and a method to construct input based on a token list.

Add IgnoreJson property to neighbours attributes as front-end does not need it and it increases data size.

Signed-off-by: Edan Toledo <TLDEDA001@myuct.ac.za>