Add HashSet to optimize longestPath algorithm.
- Better performance of algorithm.
- Better User experience.

Create HashSet that records all node ID that are already in a longest path. Once a node ID in the HashSet has its turn to be the start node, it is skipped as a longer path cannot be found since it is already in the middle of a longest path.